+++ blackbox.php
@@ -1,0 +1,30 @@
+<?php
+declare(strict_types = 1);
+
+require 'vendor/autoload.php';
+
+use Innmind\BlackBox\{
+    Application,
+    Runner\Load,
+    Runner\CodeCoverage,
+};
+
+// because the generated trees can be quite large
+\ini_set('memory_limit', '-1');
+
+Application::new($argv)
+    ->codeCoverage(
+        CodeCoverage::of(
+            __DIR__.'/src/',
+            __DIR__.'/proofs/',
+            __DIR__.'/fixtures/',
+        )
+            ->dumpTo('coverage.clover')
+            ->enableWhen(\getenv('ENABLE_COVERAGE') !== false),
+    )
+    ->scenariiPerProof(match (\getenv('ENABLE_COVERAGE')) {
+        false => 20,
+        default => 1,
+    })
+    ->tryToProve(Load::everythingIn(__DIR__.'/proofs/'))
+    ->exit();