+++ properties/Adapter/AddDirectoryFromAnotherAdapterWithFileRemoved.php
@@ -4,14 +4,25 @@ declare(strict_types = 1);
 namespace Properties\Innmind\Filesystem\Adapter;
 
 use Innmind\Filesystem\{
+    Adapter,
     Directory\Directory,
     File,
     Name,
 };
 use Innmind\Immutable\Set;
-use Innmind\BlackBox\Property;
-use PHPUnit\Framework\Assert;
+use Innmind\BlackBox\{
+    Property,
+    Set as DataSet,
+    Runner\Assert,
+};
+use Fixtures\Innmind\Filesystem\{
+    Name as FName,
+    File as FFile,
+};
 
+/**
+ * @implements Property<Adapter>
+ */
 final class AddDirectoryFromAnotherAdapterWithFileRemoved implements Property
 {
     private Name $name;
@@ -25,9 +36,14 @@ final class AddDirectoryFromAnotherAdapterWithFileRemoved implements Property
         $this->removed = $removed;
     }
 
-    public function name(): string
+    public static function any(): DataSet
     {
-        return "Add directory '{$this->name->toString()}' loaded from another adapter with file '{$this->removed->name()->toString()}' removed";
+        return DataSet\Composite::immutable(
+            static fn(...$args) => new self(...$args),
+            FName::any(),
+            FFile::any(),
+            FFile::any(),
+        );
     }
 
     public function applicableTo(object $adapter): bool
@@ -35,7 +51,7 @@ final class AddDirectoryFromAnotherAdapterWithFileRemoved implements Property
         return !$adapter->contains($this->name);
     }
 
-    public function ensureHeldBy(object $adapter): object
+    public function ensureHeldBy(Assert $assert, object $adapter): object
     {
         // directories loaded from other adapters have files injecting at
         // construct time (so there is no modifications())
@@ -48,10 +64,10 @@ final class AddDirectoryFromAnotherAdapterWithFileRemoved implements Property
         );
         $directory = $directory->remove($this->removed->name());
 
-        Assert::assertFalse($adapter->contains($directory->name()));
-        Assert::assertNull($adapter->add($directory));
-        Assert::assertTrue($adapter->contains($directory->name()));
-        Assert::assertTrue(
+        $assert->false($adapter->contains($directory->name()));
+        $assert->null($adapter->add($directory));
+        $assert->true($adapter->contains($directory->name()));
+        $assert->true(
             $adapter
                 ->get($directory->name())
                 ->match(
@@ -59,7 +75,7 @@ final class AddDirectoryFromAnotherAdapterWithFileRemoved implements Property
                     static fn() => false,
                 ),
         );
-        Assert::assertFalse(
+        $assert->false(
             $adapter
                 ->get($directory->name())
                 ->match(
@@ -67,7 +83,7 @@ final class AddDirectoryFromAnotherAdapterWithFileRemoved implements Property
                     static fn() => true,
                 ),
         );
-        Assert::assertSame(
+        $assert->same(
             $this->file->content()->toString(),
             $adapter
                 ->get($directory->name())