+++ properties/Adapter/AddEmptyDirectory.php
@@ -4,12 +4,20 @@ declare(strict_types = 1);
 namespace Properties\Innmind\Filesystem\Adapter;
 
 use Innmind\Filesystem\{
+    Adapter,
     Directory\Directory,
     Name,
 };
-use Innmind\BlackBox\Property;
-use PHPUnit\Framework\Assert;
+use Innmind\BlackBox\{
+    Property,
+    Set,
+    Runner\Assert,
+};
+use Fixtures\Innmind\Filesystem\Name as FName;
 
+/**
+ * @implements Property<Adapter>
+ */
 final class AddEmptyDirectory implements Property
 {
     private Name $name;
@@ -19,9 +27,9 @@ final class AddEmptyDirectory implements Property
         $this->name = $name;
     }
 
-    public function name(): string
+    public static function any(): Set
     {
-        return "Add empty directory '{$this->name->toString()}'";
+        return FName::any()->map(static fn($name) => new self($name));
     }
 
     public function applicableTo(object $adapter): bool
@@ -29,14 +37,14 @@ final class AddEmptyDirectory implements Property
         return !$adapter->contains($this->name);
     }
 
-    public function ensureHeldBy(object $adapter): object
+    public function ensureHeldBy(Assert $assert, object $adapter): object
     {
         $directory = Directory::of($this->name);
 
-        Assert::assertFalse($adapter->contains($directory->name()));
-        Assert::assertNull($adapter->add($directory));
-        Assert::assertTrue($adapter->contains($directory->name()));
-        Assert::assertSame(
+        $assert->false($adapter->contains($directory->name()));
+        $assert->null($adapter->add($directory));
+        $assert->true($adapter->contains($directory->name()));
+        $assert->same(
             [],
             $adapter->get($directory->name())->match(
                 static fn($dir) => $dir->reduce(