+++ properties/Directory/AllFilesAreAccessible.php
@@ -3,14 +3,21 @@ declare(strict_types = 1);
 
 namespace Properties\Innmind\Filesystem\Directory;
 
-use Innmind\BlackBox\Property;
-use PHPUnit\Framework\Assert;
+use Innmind\Filesystem\Directory;
+use Innmind\BlackBox\{
+    Property,
+    Set,
+    Runner\Assert,
+};
 
+/**
+ * @implements Property<Directory>
+ */
 final class AllFilesAreAccessible implements Property
 {
-    public function name(): string
+    public static function any(): Set
     {
-        return 'All files are accessible';
+        return Set\Elements::of(new self);
     }
 
     public function applicableTo(object $directory): bool
@@ -18,9 +25,9 @@ final class AllFilesAreAccessible implements Property
         return true;
     }
 
-    public function ensureHeldBy(object $directory): object
+    public function ensureHeldBy(Assert $assert, object $directory): object
     {
-        Assert::assertSame(
+        $assert->same(
             $directory->reduce(
                 [],
                 static fn($all, $file) => \array_merge($all, [$file]),