+++ properties/Directory/FlatMapFiles.php
@@ -5,14 +5,21 @@ namespace Properties\Innmind\Filesystem\Directory;
 
 use Innmind\Filesystem\{
     File,
-    Directory\Directory,
+    Directory,
     Name,
 };
 use Innmind\Immutable\Set;
-use Innmind\BlackBox\Property;
-use PHPUnit\Framework\Assert;
+use Innmind\BlackBox\{
+    Property,
+    Set as DataSet,
+    Runner\Assert,
+};
+use Fixtures\Innmind\Filesystem\File as FFile;
 use Ramsey\Uuid\Uuid;
 
+/**
+ * @implements Property<Directory>
+ */
 final class FlatMapFiles implements Property
 {
     private File $file1;
@@ -24,9 +31,13 @@ final class FlatMapFiles implements Property
         $this->file2 = $file2;
     }
 
-    public function name(): string
+    public static function any(): DataSet
     {
-        return 'FlatMap files';
+        return DataSet\Composite::immutable(
+            static fn(...$args) => new self(...$args),
+            DataSet\Randomize::of(FFile::any()),
+            DataSet\Randomize::of(FFile::any()),
+        );
     }
 
     public function applicableTo(object $directory): bool
@@ -34,10 +45,10 @@ final class FlatMapFiles implements Property
         return !$directory->files()->empty();
     }
 
-    public function ensureHeldBy(object $directory): object
+    public function ensureHeldBy(Assert $assert, object $directory): object
     {
         // we use uuids to avoid duplicates
-        $directory2 = $directory->flatMap(fn($file) => Directory::of(
+        $directory2 = $directory->flatMap(fn($file) => Directory\Directory::of(
             Name::of('doesntmatter'),
             Set::of(
                 $this->file1->rename(Name::of(Uuid::uuid4()->toString())),
@@ -45,11 +56,17 @@ final class FlatMapFiles implements Property
             ),
         ));
 
-        Assert::assertNotSame($directory, $directory2);
-        Assert::assertSame($directory->name()->toString(), $directory2->name()->toString());
-        Assert::assertNotSame($directory->files(), $directory2->files());
-        Assert::assertSame($directory->files()->size() * 2, $directory2->files()->size());
-        Assert::assertSame(
+        $assert
+            ->expected($directory)
+            ->not()
+            ->same($directory2);
+        $assert->same($directory->name()->toString(), $directory2->name()->toString());
+        $assert
+            ->expected($directory->files())
+            ->not()
+            ->same($directory2->files());
+        $assert->same($directory->files()->size() * 2, $directory2->files()->size());
+        $assert->same(
             [$this->file1->content(), $this->file2->content()],
             $directory2
                 ->files()