+++ properties/Directory/MediaTypeIsAlwaysTheSame.php
@@ -3,14 +3,21 @@ declare(strict_types = 1);
 
 namespace Properties\Innmind\Filesystem\Directory;
 
-use Innmind\BlackBox\Property;
-use PHPUnit\Framework\Assert;
+use Innmind\Filesystem\Directory;
+use Innmind\BlackBox\{
+    Property,
+    Set,
+    Runner\Assert,
+};
 
+/**
+ * @implements Property<Directory>
+ */
 final class MediaTypeIsAlwaysTheSame implements Property
 {
-    public function name(): string
+    public static function any(): Set
     {
-        return 'Directory media type is always the same';
+        return Set\Elements::of(new self);
     }
 
     public function applicableTo(object $directory): bool
@@ -18,9 +25,9 @@ final class MediaTypeIsAlwaysTheSame implements Property
         return true;
     }
 
-    public function ensureHeldBy(object $directory): object
+    public function ensureHeldBy(Assert $assert, object $directory): object
     {
-        Assert::assertSame(
+        $assert->same(
             'text/directory',
             $directory->mediaType()->toString(),
         );