+++ tests/NameTest.php
@@ -141,8 +141,7 @@ class NameTest extends TestCase
                                 ...\range(48, 127),
                             ),
                         ),
-                        Set\Integers::between(255, 1024), // upper limit at 1024 to avoid out of memory
-                    ),
+                    )->between(255, 1024), // upper limit at 1024 to avoid out of memory
                 )->filter(static fn(string $name): bool => $name !== '.' && $name !== '..'),
             )
             ->then(function($name) {